+++ include/quill/backend/BackendManager.h
@@ -23,6 +23,12 @@ class BackendTscClock;
 
 namespace detail
 {
+
+#if defined(_WIN32) && defined(_MSC_VER) && !defined(__GNUC__)
+  #pragma warning(push)
+  #pragma warning(disable : 4324)
+#endif
+
 /**
  * Provides access to common collection class that are used by both the frontend and the backend
  * components of the logging system
@@ -102,10 +108,29 @@ private:
   }
 
 private:
+  /***/
+  QUILL_NODISCARD bool is_atexit_registered() const noexcept
+  {
+    return _atexit_registered.load();
+  }
+
+  /***/
+  void set_atexit_registered() noexcept
+  {
+    _atexit_registered.store(true);
+  }
+
+private:
   BackendWorker _backend_worker;
   ManualBackendWorker _manual_backend_worker{&_backend_worker};
   std::atomic<std::once_flag*> _start_once_flag{new std::once_flag};
+  std::atomic<bool> _atexit_registered{false};
 };
+
+#if defined(_WIN32) && defined(_MSC_VER) && !defined(__GNUC__)
+  #pragma warning(pop)
+#endif
+
 } // namespace detail
 
 QUILL_END_NAMESPACE